pyproject.toml
[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core"]

[tool.poetry]
authors = [
  "Moritz Ulmer <moritz@silentwind.eu>",
]
classifiers = [
  "Operating System :: OS Independent",
  "Programming Language :: Python :: 3.10",
  "Operating System :: Microsoft :: Windows",
  "Operating System :: POSIX",
  "License :: OSI Approved :: Apache Software License",
]
description = "A microcontroller flasher for the Inamata IoT platform"
include = ["fonts/*", "images/*", "uis/*", "translations/*", "littlefs_partition/*"]
license = "Apache-2.0"
name = "inamata_flasher"
packages = [{include = "esp_idf/..", from = "src"}]
readme = "README.md"
version = "1.3.0"

[tool.poetry.dependencies]
PySide6 = "~=6.6.0"
platformdirs = "~=4.0.0"
esptool = "~=4.6.2"
keyring = "~=24.3.0"
littlefs-python = "~=0.4.0"
pyjwt = {version = "~=2.8.0", extras = ["crypto"]}
pyserial = "~=3.5"
python = "~=3.10.0"
requests = "~=2.31.0"
semantic-version = "~=2.10.0"

[tool.poetry.group.dev.dependencies]
black = "~=23.3"
debugpy = "~=1.6"
ipdb = "0.13.13"
pip-check = "~=2.8"
pyinstaller = "~=6.2"
# Used for linting tests
ruff = "~=0.1.6"
# Automate build process
pybuilder = "^0.13.10"

[tool.poetry.scripts]
inamata-flasher = "main:main"

[tool.ruff]
src = ["src"]
# Flake8 F, pycodestyle E and W, mccabe complexity C90, isort I
select = ["F", "E", "W", "C90", "I"]
# E501: line too long
ignore = ["E501"]
# Package files, system libs
extend-exclude = [
  ".cache/pypoetry/virtualenvs/",
  "/usr/lib/",
]

# Assume Python 3.10.
target-version = "py310"

[tool.ruff.mccabe]
max-complexity = 13
